refactor(MapTopology): relocate MapTopology out of ai/ to services/
Moves MapTopology.ts from src/server/services/ai/ to src/server/services/
so general-purpose hex/grid utilities are no longer falsely scoped to
AI-only consumers. Updates all 24 importers (10 production files, 14 test
files). Also relocates ai/MapTopology.test.ts to MapTopology.ai.test.ts
to preserve both test files. Build and full test suite (2072 tests) pass
with no regressions.
