+++ src/server/__tests__/ai/AIStrategyEngine.test.ts
@@ -27,7 +27,7 @@ jest.mock('../../services/socketService', () => ({
   getSocketIO: jest.fn<() => any>().mockReturnValue(null),
 }));
 
-jest.mock('../../services/ai/MapTopology', () => ({
+jest.mock('../../services/MapTopology', () => ({
   loadGridPoints: jest.fn(() => new Map()),
   getHexNeighbors: jest.fn(() => []),
   getTerrainCost: jest.fn(() => 1),
@@ -198,7 +198,7 @@ import { TurnComposer } from '../../services/ai/TurnComposer';
 import { db } from '../../db/index';
 import { emitToGame } from '../../services/socketService';
 import { getMemory, updateMemory } from '../../services/ai/BotMemory';
-import { loadGridPoints } from '../../services/ai/MapTopology';
+import { loadGridPoints } from '../../services/MapTopology';
 import { PlayerService } from '../../services/playerService';
 import {
   AIActionType,
@@ -4082,7 +4082,7 @@ describe('AIStrategyEngine.takeTurn (Integration)', () => {
       mockFindDeadLoads.mockReturnValue(deadLoads);
 
       // Mock loadGridPointsMap to return a city at bot's position
-      const { loadGridPoints } = require('../../services/ai/MapTopology');
+      const { loadGridPoints } = require('../../services/MapTopology');
       (loadGridPoints as jest.Mock).mockReturnValue(new Map([
         ['10,10', { name: 'TestCity' }],
       ]));
@@ -4160,7 +4160,7 @@ describe('AIStrategyEngine.takeTurn (Integration)', () => {
       mockFindDeadLoads.mockReturnValue(['Hops']);
 
       // loadGridPointsMap returns no city at bot's position
-      const { loadGridPoints } = require('../../services/ai/MapTopology');
+      const { loadGridPoints } = require('../../services/MapTopology');
       (loadGridPoints as jest.Mock).mockReturnValue(new Map());
 
       const route = {
