+++ src/server/__tests__/ai/ActionResolver.test.ts
@@ -30,14 +30,14 @@ import {
   TurnPlanDiscardHand,
   TurnPlanPassTurn,
 } from '../../../shared/types/GameTypes';
-import type { GridPointData, GridCoord } from '../../services/ai/MapTopology';
+import type { GridPointData, GridCoord } from '../../services/MapTopology';
 import type { TrackUsageComputation, PathEdge } from '../../../shared/services/trackUsageFees';
 
 // ─── Mock modules ────────────────────────────────────────────────────────────
 
 jest.mock('../../services/ai/computeBuildSegments');
 jest.mock('../../../shared/services/trackUsageFees');
-jest.mock('../../services/ai/MapTopology');
+jest.mock('../../services/MapTopology');
 jest.mock('../../../shared/services/majorCityGroups', () => {
   const actual = jest.requireActual('../../../shared/services/majorCityGroups');
   return {
@@ -50,7 +50,7 @@ jest.mock('../../../shared/services/majorCityGroups', () => {
 
 import { computeBuildSegments } from '../../services/ai/computeBuildSegments';
 import { computeTrackUsageForMove } from '../../../shared/services/trackUsageFees';
-import { loadGridPoints, hexDistance } from '../../services/ai/MapTopology';
+import { loadGridPoints, hexDistance } from '../../services/MapTopology';
 import { getMajorCityGroups, getMajorCityLookup } from '../../../shared/services/majorCityGroups';
 
 const mockComputeBuildSegments = computeBuildSegments as jest.MockedFunction<typeof computeBuildSegments>;
