+++ src/server/__tests__/ai/WorldSnapshotService.test.ts
@@ -14,7 +14,7 @@ jest.mock('../../db/index', () => ({
   db: { query: jest.fn() },
 }));
 
-jest.mock('../../services/ai/MapTopology', () => ({
+jest.mock('../../services/MapTopology', () => ({
   loadGridPoints: jest.fn(() => new Map()),
   gridToPixel: jest.fn(() => ({ x: 0, y: 0 })),
 }));
@@ -47,7 +47,7 @@ jest.mock('../../services/loadService', () => ({
 
 import { capture } from '../../services/ai/WorldSnapshotService';
 import { db } from '../../db/index';
-import { loadGridPoints } from '../../services/ai/MapTopology';
+import { loadGridPoints } from '../../services/MapTopology';
 
 const mockQuery = db.query as jest.Mock;
 const mockLoadGridPoints = loadGridPoints as jest.Mock;
