+++ src/server/services/ai/AIStrategyEngine.ts
@@ -41,7 +41,7 @@ import {
 import { db } from '../../db/index';
 import { getMajorCityGroups, getMajorCityLookup, computeEffectivePathLength } from '../../../shared/services/majorCityGroups';
 import { getTrainCapacity } from '../../../shared/services/trainProperties';
-import { gridToPixel, loadGridPoints as loadGridPointsMap } from './MapTopology';
+import { gridToPixel, loadGridPoints as loadGridPointsMap } from '../MapTopology';
 import { RouteValidator } from './RouteValidator';
 import { getMemory, updateMemory } from './BotMemory';
 import { initTurnLog, logPhase, flushTurnLog, LLMPhaseFields } from './DecisionLogger';
