+++ src/server/services/ai/ActionResolver.ts
@@ -28,7 +28,7 @@ import {
   TRACK_USAGE_FEE,
   PlayerTrackState,
 } from '../../../shared/types/GameTypes';
-import { loadGridPoints, GridCoord, GridPointData, hexDistance } from './MapTopology';
+import { loadGridPoints, GridCoord, GridPointData, hexDistance } from '../MapTopology';
 import { getMajorCityGroups, getMajorCityLookup, computeEffectivePathLength } from '../../../shared/services/majorCityGroups';
 import { computeBuildSegments } from './computeBuildSegments';
 import { computeTrackUsageForMove } from '../../../shared/services/trackUsageFees';
