+++ src/server/services/ai/TurnExecutor.ts
@@ -18,7 +18,7 @@ import { emitToGame, emitStatePatch } from '../socketService';
 import { PlayerService } from '../playerService';
 import { LoadService } from '../loadService';
 import { DemandDeckService } from '../demandDeckService';
-import { gridToPixel, loadGridPoints } from './MapTopology';
+import { gridToPixel, loadGridPoints } from '../MapTopology';
 import { getTrainCapacity, getTrainSpeed } from '../../../shared/services/trainProperties';
 import { getCityNameAtPosition } from '../../../shared/services/cityPositionResolver';
